+++ plugins/renovate-backend/src/config/index.ts
@@ -1,8 +1,8 @@
 import { Config } from '@backstage/config';
 import {
-  readTaskScheduleDefinitionFromConfig,
-  TaskScheduleDefinition,
-} from '@backstage/backend-tasks';
+  readSchedulerServiceTaskScheduleDefinitionFromConfig,
+  SchedulerServiceTaskScheduleDefinition,
+} from '@backstage/backend-plugin-api';
 import { JsonValue } from '@backstage/types';
 
 export const RENOVATE_ANNOTATION_KEEP_UPDATED =
@@ -31,11 +31,11 @@ export function getRuntimeConfigs(rootConfig: Config): {
 
 export function getScheduleDefinition(
   pluginConfig: Config,
-  variant: 'jobSync' | 'renovation' | 'cleanup',
-): TaskScheduleDefinition {
+  variant: 'renovation' | 'cleanup',
+): SchedulerServiceTaskScheduleDefinition {
   try {
     const scheduleConfig = pluginConfig.getConfig(`schedules.${variant}`);
-    return readTaskScheduleDefinitionFromConfig(scheduleConfig);
+    return readSchedulerServiceTaskScheduleDefinitionFromConfig(scheduleConfig);
   } catch (e) {
     return {
       scope: 'global',