+++ plugins/renovate-backend/src/wrapper/types.ts
@@ -5,7 +5,3 @@ export interface ExtractReportOptions {
   logStream: Readable;
   logger: LoggerService;
 }
-
-export interface RunOptions {
-  logger: LoggerService;
-}