+++ plugins/renovate-node/src/index.ts
@@ -1,2 +1,4 @@
 export type * from './extensionPoints';
 export * from './extensionPoints';
+export type * from './queue/types';
+export type * from './types';