+++ plugins/renovate-node/src/types.ts
@@ -1,0 +1,6 @@
+import { TargetRepo } from '@secustor/backstage-plugin-renovate-common';
+
+export interface RunOptions {
+  id: string;
+  target: TargetRepo;
+}